factor torch out
