+++ models/Models.py
@@ -180,30 +180,8 @@ def make_mnist_model(**args):
     model.add(Activation('softmax'))
     return model
 
-class MNistNet(torch.nn.Module):
-    def __init__(self):
-        super(MNistNet, self).__init__()
-        self.conv1 = nn.Conv2d(1, 10, kernel_size=5)
-        self.conv2 = nn.Conv2d(10, 20, kernel_size=5)
-        self.conv2_drop = nn.Dropout2d()
-        self.fc1 = nn.Linear(320, 50)
-        self.fc2 = nn.Linear(50, 10)
-        
-    def forward(self, x):
-        x = x.permute(0,3,1,2).float()
-        x = F.relu(F.max_pool2d(self.conv1(x), 2))
-        x = F.relu(F.max_pool2d(self.conv2_drop(self.conv2(x)), 2))
-        x = x.view(-1, 320)
-        x = F.relu(self.fc1(x))
-        x = F.dropout(x, training=self.training)
-        x = self.fc2(x)
-        return F.log_softmax(x, dim=1)
-        #return F.softmax(x)
-        #return F.cross_entropy(x)
-        #return x
-        
-            
 def make_mnist_torch_model(**args):
+    from PytorchCNN import MNistNet
     model = MNistNet()
     return model
